Fixed memory issues with kernel parameters
I made some calculation issues originally since I based it of 0x88000000
insetad of 0x87300000, so it overrode camera memory. However, when
fixing this, then it overrode dsplink memory. Thus, dsplink must occupy
the memory after cmem like it did with DVSDK. It appears to work with
13MiB and not with 11 or 12, so I'd guess it's 13 MiB. However, I was
unable to determine this from the code (hence the TODO comment in the
README).