Added TIImgenc1 patches for multiple images
Anything after the first image saved using TIImgenc1 is a 575 jpeg with
no useful data in it. For example, run this and look at the fakesink
output. You should see the first image be something like 10670 bytes and
the rest 575 bytes.

    gst-launch-0.10 v4l2src device=/dev/video2 -v num-buffers=3 ! \
    video/x-raw-yuv,format=\(fourcc\)UYVY, \
    framerate=\(fraction\)30000/1001,width=640,height=480 ! \
    TIImgenc1 engineName=codecServer codecName=jpegenc \
    iColorSpace=UYVY oColorSpace=YUV420P qValue=75 numOutputBufs=3 \
    genTimeStamps=1 ! fakesink

The patches applied were described in the following link and should
resolve this issue.

http://e2e.ti.com/support/dsp/omap_applications_processors/f/447/t/138400.aspx