cmem memory modifications for full-res pools
This will support having two pools for 2592x1944 images from the camera.
These changes aren't needed for use with video, but they don't hurt
unless you really need the extra 44 MiB of RAM. Also just completely
remove the kernel argument checks in the gstti-init service so we don't
have to mess with that every time we mess with cmem.