Fixed blank license file, added random URL
We're required to have this file even for pointless "packages" that just
install a service file. Thus, it's blank. However, I forgot to include
it in the package...

I also added a URL to jumpnowtek's tutorial about USB Networking for
the 2.6.x kernel, which sort of is useful. We're also required to have a
URL even for "packages" that don't have a website.