Install libgles and omap3-sgx-modules 4.09.00.01
The 4.05 version fails with "Unknown symbol OMAPLFBGetUpdateMode" due to a
patch in meta-ti removing it, but if we don't remove it, then we have to
revert the OMAP DSS removal patch as described in the below link. In the end,
just using 4.09 works, and is a whole lot simpler to get working.

https://github.com/scottellis/sgx-qt-build