Added dev packages for use with OpenGL ES
These are needed to compile stuff with libgles{1,2} (e.g. `g++ file.cpp -lEGL -lGLES_CM'
or `g++ file.cpp -lEGL -lGLESv2'). Note that with version 2.0, I can't seem to create a
context, so I have had to resort to working with version 1.1 at the moment. Same problem
as described here:

http://forum.imgtec.com/discussion/3148/getting-error-on-eglcreatecontext-with-error-loadogles2andgetfunctions-couldn-t-get-function-table

Also, the package gst-plugins-gl does not work yet, but it is closer to working when
GLEW is listed as a dependency as described below.

http://patchwork.openembedded.org/patch/55603/